refactor: file extensions and location
