+++ src/app/analise/page.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/app/components/Navbar";
+import Navbar from "@/components/customComponents/Navbar";
 import Image from "next/image";
 import { Checkbox } from "@/components/ui/checkbox"
 import { Button } from "@/components/ui/button";
@@ -85,12 +85,12 @@ export default function Analise() {
                     </div>
 
                     <div className="w-full flex justify-between items-center mt-7 border-t-1 border-t-gray-300 pb-3 pt-2">
-                        <Link href="/pages/dashboard">
+                        <Link href="/dashboard">
                         <Button className="bg-white text-[var(--color-azul-escuro)] border-1 border-gray-300">
                             Cancelar
                         </Button>
                         </Link>
-                        <Link href="/pages/analisedocs"><Button>Próximo</Button></Link>
+                        <Link href="/analisedocs"><Button>Próximo</Button></Link>
                     </div>
                 </div>
                 <div className="flex-1" />
