+++ src/app/analiseinfo/page.tsx
@@ -1,5 +1,5 @@
 "use client"
-import Navbar from "@/app/components/Navbar";
+import Navbar from "@/components/customComponents/Navbar";
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label"
 import { Input } from "@/components/ui/input"
@@ -118,7 +118,7 @@ export default function AnaliseInfo() {
                     </div>
 
                     <div className="w-full flex justify-between items-center mt-7 border-t-1 border-t-gray-300 pb-3 pt-2">
-                        <Link href="/pages/analise">
+                        <Link href="/analise">
                             <Button className="bg-white text-[var(--color-azul-escuro)] border-1 border-gray-300">
                                 Cancelar
                             </Button>
