+++ src/app/page.tsx
@@ -45,10 +45,10 @@ const Login: React.FC = () => {
 
   return (
     <form onSubmit={login}>
-      <div className="w-sm h-xs border-0 border-gray-300 rounded-lg px-8 pt-8 pb-11"
+      <div className= "w-screen h-auto px-4 border-0 border-gray-300 rounded-lg xs:w-sm xs:h-xs mt-6 xt:mt-8 2xl:mt-10"
         style={{ boxShadow: '0px 4px 6px 0px rgba(0, 0, 0, 0.10), 0px 10px 15px 0px rgba(0, 0, 0, 0.10)' }}>
 
-        <div className="grid w-full max-w-sm items-center gap-1.5 relative">
+        <div className="grid w-full max-w-sm items-center gap-1.5 relative xt:mt-8">
           <Label htmlFor="email" className="font-normal">Email</Label>
           <Image
             src="/email.svg"
